siphon-server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "siphon_server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12, <3.14"
dependencies = [
    "conduit",
    "dbclients",
    "fastapi>=0.120.0",
    "markitdown[all]",
    "rich>=14.2.0",
    "siphon_api",
    "pydub",
    "sqlalchemy>=2.0.44",
    "torch",
    "trafilatura[all]>=2.0.0",
    "xdg-base-dirs>=6.0.2",
    "youtube-transcript-api>=1.2.3",
    "yt-dlp>=2025.08.01",
]

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true

[tool.hatch.build.targets.wheel]
packages = ["src/siphon_server"]
sources = ["src"]
include = ["src/siphon_server/py.typed"]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }
siphon_api = { path = "../siphon-api", editable = true }
conduit = { path = "../../conduit-project", editable = true }
dbclients = { path = "../../dbclients-project", editable = true }

[dependency-groups]
dev = [
    "pytest-watcher>=0.4.3",
]

[project.scripts]
siph = "siphon_server.cli.siphon_cli:main"
diarization_service= "siphon_server.workers.diarization_cpu.launcher:main"
